Reworked entire structure and handling of TransactionsTable + management of tanstack/query + paginated-data-table to work half-correctly with Suspense.
