+++ src/app/transactions/page.tsx
@@ -1,12 +1,41 @@
-import TransactionsTable from "@/components/TransactionsTable";
+import { TransactionsTable } from "@/components/TransactionsTable";
+import getTransactions from "@/components/TransactionsTable/getTransactions";
+import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query";
 
+// TODO: do we want this anymore? what is the story of caching between the client and events.
 export const dynamic = "force-dynamic";
 
-const Page = async () => {
+const Page = () => {
+  const queryClient = new QueryClient();
+
+  const defaultQueryOptions = {
+    pageIndex: 0,
+    pageSize: 10,
+  };
+
+  const endpoint = "/api/transactions";
+  const queryName = "TransactionsTable";
+  const errorMessage = "Failed to query data while trying to generate event table, please try reloading the page.";
+
+  queryClient.prefetchQuery({
+    queryFn: () => getTransactions({ endpoint, pageIndex: 0}),
+    queryKey: [queryName, defaultQueryOptions],
+    meta: {
+      errorMessage,
+    },
+  });
+
   return (
     <div className="bg-primary flex flex-col gap-5 pt-5 items-center">
       <h1 className="sm:text-2xl font-bold">Recent Transactions</h1>
-      <TransactionsTable className="sm:w-11/12 w-full"/>
+      <HydrationBoundary state={dehydrate(queryClient)}>
+        <TransactionsTable
+          className="sm:w-11/12 w-full"
+          queryName={queryName}
+          defaultQueryOptions={defaultQueryOptions}
+          endpoint={endpoint}
+          errorMessage={errorMessage}/>
+      </HydrationBoundary>
     </div>
   );
 };
