+++ src/components/ui/paginated-data-table.tsx
@@ -1,177 +1,78 @@
-"use client";
+import { ColumnDef, Table as TableT, flexRender } from "@tanstack/react-table";
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "./Table";
+import { Button } from "./button";
 
-import {
-  type ColumnDef,
-  flexRender,
-  getCoreRowModel,
-  useReactTable,
-  type PaginationState,
-} from "@tanstack/react-table";
-
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/Table";
-
-import { Button } from "@/components/ui/button";
-import { useMemo, useState } from "react";
-import { useQuery } from "@tanstack/react-query";
-import { type z } from "zod";
-
-export interface QueryOptions {
-  pageIndex: number,
-  pageSize: number,
-}
-
-interface FetcherParams {
-  endpoint: string,
-  pageIndex: number
-};
-
-type Fetcher<Z extends z.ZodTypeAny> = ({ endpoint, pageIndex} : FetcherParams) => Promise<z.infer<Z>>;
-
-interface PaginatedDataTableProps<TData, TValue, Z extends z.ZodTypeAny> {
-  className?: string,
+interface PaginatedDataTableProps<TData, TValue> {
+  table: TableT<TData>,
   columns: Array<ColumnDef<TData, TValue>>
-  queryName: string,
-  defaultQueryOptions: QueryOptions,
-  endpoint: string,
-  fetcher: Fetcher<Z>,
-  errorMessage: string,
 }
 
-export function PaginatedDataTable<TData, TValue, Z extends z.ZodTypeAny>({
-  className,
-  columns,
-  queryName,
-  defaultQueryOptions,
-  endpoint,
-  fetcher,
-  errorMessage,
-}: PaginatedDataTableProps<TData, TValue, Z>) {
+export function PaginatedDataTable<TData, TValue> ({ table, columns } : PaginatedDataTableProps<TData, TValue> ) {
 
-  const [{ pageIndex, pageSize }, setPagination] = useState<PaginationState>({...defaultQueryOptions});
-
-  const queryOptions = {
-    pageIndex,
-    pageSize,
-  };
-
-  const { data, isError, isPending, isSuccess } = useQuery({
-    queryKey: [queryName, queryOptions],
-    queryFn: async () => await fetcher({ endpoint, pageIndex }),
-    meta: {
-      errorMessage,
-    },
-  });
-
-  const { pages: pageCount, results: tableData } : z.infer<Z> = data ?? { pages: 0, results: []};
-
-  const pagination = useMemo(
-    () => ({
-      pageIndex,
-      pageSize,
-    }),
-    [pageIndex, pageSize],
-  );
-
-  const table = useReactTable({
-    data: tableData as TData[],
-    columns,
-    pageCount,
-    state: {
-      pagination,
-    },
-    onPaginationChange: setPagination,
-    getCoreRowModel: getCoreRowModel(),
-    manualPagination: true,
-  });
-
-  if (isSuccess){
-    return (
-      <div className={`${className ?? ""}`}>
-        <div className="rounded-md border">
-          <Table>
-            <TableHeader>
-              {table.getHeaderGroups().map((headerGroup) => (
-                <TableRow key={headerGroup.id}>
-                  {headerGroup.headers.map((header) => {
-                    return (
-                      <TableHead key={header.id}>
-                        {header.isPlaceholder
-                          ? null
-                          : flexRender(
-                              header.column.columnDef.header,
-                              header.getContext(),
-                            )}
-                      </TableHead>
-                    );
-                  })}
-                </TableRow>
-              ))}
-            </TableHeader>
-            <TableBody>
-              {table.getRowModel().rows?.length ? (
-                table.getRowModel().rows.map((row) => (
-                  <TableRow
-                    key={row.id}
-                    data-state={row.getIsSelected() && "selected"}
-                  >
-                    {row.getVisibleCells().map((cell) => (
-                      <TableCell key={cell.id}>
-                        {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                      </TableCell>
-                    ))}
-                  </TableRow>
-                ))
-              ) : (
-                <TableRow>
-                  <TableCell colSpan={columns.length} className="h-24 text-center">
-                    No results.
+  return (
+    <div>
+    <div className="rounded-md border">
+      <Table>
+        <TableHeader>
+          {table.getHeaderGroups().map((headerGroup) => (
+            <TableRow key={headerGroup.id}>
+              {headerGroup.headers.map((header) => {
+                return (
+                  <TableHead key={header.id}>
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(
+                          header.column.columnDef.header,
+                          header.getContext(),
+                        )}
+                  </TableHead>
+                );
+              })}
+            </TableRow>
+          ))}
+        </TableHeader>
+        <TableBody>
+          {table.getRowModel().rows?.length ? (
+            table.getRowModel().rows.map((row) => (
+              <TableRow
+                key={row.id}
+                data-state={row.getIsSelected() && "selected"}
+              >
+                {row.getVisibleCells().map((cell) => (
+                  <TableCell key={cell.id}>
+                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
                   </TableCell>
-                </TableRow>
-              )}
-            </TableBody>
-          </Table>
-        </div>
-        <div className="flex items-center justify-end space-x-2 py-4">
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => {table.previousPage()}}
-            disabled={!table.getCanPreviousPage()}
-          >
-            Previous
-          </Button>
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => {table.nextPage()}}
-            disabled={!table.getCanNextPage()}
-          >
-            Next
-          </Button>
-        </div>
-       </div>
-    );
-  }
-  if (isPending) {
-    return (
-      <div className="rounded-sm shadow-md">
-        <h1 className="text-lg font-semibold">Loading...</h1>
-      </div>
-    );
-  }
-
-  if (isError) {
-    return (
-      <div className="rounded-sm shadow-md py-5 flex justify-center">
-        <h1 className="text-lg font-semibold">No results found.</h1>
-      </div>
-    );
-  }
+                ))}
+              </TableRow>
+            ))
+          ) : (
+            <TableRow>
+              <TableCell colSpan={columns.length} className="h-24 text-center">
+                No results.
+              </TableCell>
+            </TableRow>
+          )}
+        </TableBody>
+      </Table>
+    </div>
+    <div className="flex items-center justify-end space-x-2 py-4">
+      <Button
+        variant="ghost"
+        size="sm"
+        onClick={() => {table.previousPage()}}
+        disabled={!table.getCanPreviousPage()}
+      >
+        Previous
+      </Button>
+      <Button
+        variant="ghost"
+        size="sm"
+        onClick={() => {table.nextPage()}}
+        disabled={!table.getCanNextPage()}
+      >
+        Next
+      </Button>
+    </div>
+   </div>
+  );
 }
